+++ apps/api/src/domains/users/users.module.ts
@@ -1,12 +1,11 @@
 import { Module } from "@nestjs/common"
 import { TypeOrmModule } from "@nestjs/typeorm"
-import { RbacModule } from "@/domains/rbac/rbac.module"
 import { User } from "./user.entity"
 import { UserRepository } from "./user.repository"
 import { UsersService } from "./users.service"
 
 @Module({
-  imports: [TypeOrmModule.forFeature([User]), RbacModule],
+  imports: [TypeOrmModule.forFeature([User])],
   providers: [UsersService, UserRepository],
   exports: [UsersService, UserRepository],
 })